refactor(memory): extract sub-methods from Letta Memorize()
Extract five helper methods from the 238-line Memorize() function:
- findSimilarMemories: dedup search against archival memory
- formatMessagesAsLines: format conversation messages for prompt
- buildConsolidationPrompt: construct the LLM consolidation prompt
- applyCoreMemoryUpdates: apply core memory block updates
- applyArchivalUpdates: write archival entries, delete duplicates, append history

Memorize() is now a clean 58-line coordinator calling these sub-methods.
